Problem on greatest of three numbers
Problem Statement - FOE college wants to recognize the department which has succeeded in getting the maximum number of placements for this academic year. The
departments that have participated In the recruitment drive are CSE,ECE, MECH. Help the college find the department getting maximum placements. Check for all the
possible output given in the sample snapshot
Note : If any input is negative, the output should be "Input is Invalid". If all department has equal number of placements, the output should be •None of the department has got
the highest placement.
Sample Input 1:
• Enter the no of students placed in CSE:90
• Enter the no of students placed in ECE:45
• Enter the no of students placed in MECH:70
Sample Output 1:
• Highest placement
CSE